+++ tests/test_forest.py
@@ -14,12 +14,16 @@ CMTOINCH = 1/2.54
 SHAPE_DICT = {'PGS only': 'o', 'PGS plus': 's', 'PGS extended': 'H'}
 COL_DICT = {'wo T2DM/CVD': 'orangered', 'w T2DM': 'blueviolet',
               'w T2DM & CVD': 'limegreen'}
+ALPHA_DICT = {'wo T2DM/CVD': .4, 'w T2DM': .65, 'w T2DM & CVD': .9}
 # NOTE keep the numeric order to simplify testing
 SORT_DICT = {'AF': 0,  'CVD': 1, 'CHD': 2,
              'HF': 3, 'CVD + AF + HF': 4,'Ischaemic Stroke': 5, }
 COL_NAME='col'
 SHAPE_NAME='shape'
+ALPHA_NAME='alpha'
 POINT = 'test_cstatistic'
+STRING_COL = 'string'
+STRING_HEAD = 'test'
 UB = POINT + '_ub'
 LB = POINT + '_lb'
 GROUP='evaluated_outcome'
@@ -33,6 +37,7 @@ data1 = examples.load_forest_data()
 #  col and shape
 data1[COL_NAME] = data1.subgroup_name.map(COL_DICT)
 data1[SHAPE_NAME] = data1.model.map(SHAPE_DICT)
+data1[ALPHA_NAME] = data1.subgroup_name.map(ALPHA_DICT)
 # select a single 'study'
 data2 = data1[data1.model=='PGS only'].copy()
 
@@ -110,7 +115,7 @@ class TestPlotForest(object):
         f, ax = plt.subplots(1, figsize=(15, 15))
         _, ax = forest.plot_forest(df=data2,
                                    x_col=POINT, lb_col=LB, ub_col=UB,
-                                   s_col=SHAPE_NAME,
+                                   s_col=SHAPE_NAME, a_col=ALPHA_NAME,
                                    c_col=COL_NAME, ci_colour='black',
                                    g_col='evaluated_outcome', shape_size= 19,
                                    ci_lwd=2,
@@ -124,9 +129,11 @@ class TestPlotForest(object):
         # check the points are correct
         assert list(data2[POINT]) ==\
             [list(cl.get_offsets().data[0])[0] for cl in ax.collections]
-        # check the shape size
+        # check the shape size and alphas
         collect=ax.collections
         assert collect[0].get_sizes() == 19
+        assert list(data2[ALPHA_NAME]) == \
+                list([al.get_alpha() for al in collect])
         # get conficence interval coordinates
         lines=ax.lines
         assert list(lines[3].get_xdata()) == \
@@ -138,14 +145,17 @@ class TestPlotForest(object):
         _, ax = forest.plot_forest(df=data2,
                                    x_col=POINT, lb_col=LB, ub_col=UB,
                                    s_col=SHAPE_NAME, c_col=COL_NAME,
+                                   a_col=ALPHA_NAME,
                                    g_col='evaluated_outcome',
                                    )
         # check the points are correct
         assert list(data2[POINT]) ==\
             [list(cl.get_offsets().data[0])[0] for cl in ax.collections]
-        # check the shape size
+        # check the shape size and alphas
         collect=ax.collections
         assert collect[0].get_sizes() == 40
+        assert list(data2[ALPHA_NAME]) == \
+                list([al.get_alpha() for al in collect])
         # get conficence interval coordinates
         lines=ax.lines
         assert list(lines[9].get_xdata()) == \
@@ -159,7 +169,7 @@ class TestPlotForest(object):
         _, ax = forest.plot_forest(df=data1,
                                    x_col=POINT, lb_col=LB, ub_col=UB,
                                    s_col=SHAPE_NAME, c_col=COL_NAME,
-                                   ci_colour='black',
+                                   a_col=ALPHA_NAME, ci_colour='black',
                                    g_col='evaluated_outcome',
                                    connect_shape=True,
                                    kwargs_scatter_dict={'edgecolors':'black'},
@@ -168,14 +178,44 @@ class TestPlotForest(object):
         # check the points are correct
         assert list(data1[POINT]) ==\
             [list(cl.get_offsets().data[0])[0] for cl in ax.collections]
-        # check the shape size
+        # check the shape size and alphas
         collect=ax.collections
         assert collect[0].get_sizes() == 40
+        assert list(data1[ALPHA_NAME]) == \
+                list([al.get_alpha() for al in collect])
         # get conficence interval coordinates
         lines=ax.lines
         assert list(lines[9].get_xdata()) == \
-            list(data2.loc[:,[LB, UB]].iloc[9])
+            list(data1.loc[:,[LB, UB]].iloc[9])
         assert lines[5].get_linestyle() == '-'
         assert lines[6].get_solid_capstyle() == 'projecting'
         assert lines[3].get_lw() == 2.0
         assert lines[3].get_color() == 'black'
+
+# @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
+# plot_table
+class TestPlotTable(object):
+    """
+    Testing the `plot_table` function.
+    """
+    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    def test_plot_table(self):
+        # figure
+        _, ax = plt.subplots(1, figsize=(5,5))
+        # string
+        data2[STRING_COL] = data2[POINT].map('{:,.2f}'.format)
+        ax.set_ylim(min(data2.y_axis.to_list()), max(data2.y_axis.to_list()))
+        # the function to test
+        _ = forest.plot_table(data2, annoteheader=STRING_HEAD,
+                              string_col=STRING_COL, ax=ax,
+                              halignment_text='left',
+                              halignment_header='center',
+                              size_text=5, size_header=6,
+                              negative_padding=2,
+                              )
+        # assert
+        assert [t.get_text() for t in ax.texts] ==\
+            data2[STRING_COL].to_list() + [STRING_HEAD]
+        assert ax.texts[-1].get_fontsize() == 6
+        assert ax.texts[1].get_fontsize() == 5
+        assert ax.texts[1].get_horizontalalignment() == 'left'